+++ dbt_platform_helper/providers/version.py
@@ -1,5 +1,3 @@
-import re
-import subprocess
 from abc import ABC
 from importlib.metadata import PackageNotFoundError
 from importlib.metadata import version
@@ -10,7 +8,6 @@ import requests
 from dbt_platform_helper.constants import PLATFORM_HELPER_VERSION_FILE
 from dbt_platform_helper.platform_exception import PlatformException
 from dbt_platform_helper.providers.semantic_version import SemanticVersion
-from dbt_platform_helper.providers.semantic_version import VersionStatus
 from dbt_platform_helper.providers.yaml_file import FileProviderException
 from dbt_platform_helper.providers.yaml_file import YamlFileProvider
 
@@ -80,40 +77,3 @@ class DeprecatedVersionFileVersionProvider(VersionProvider):
         except FileProviderException:
             version_from_file = None
         return version_from_file
-
-
-class AWSVersioning:
-    def __init__(self, latest_version_provider: VersionProvider = None):
-        self.latest_version_provider = latest_version_provider or GithubLatestVersionProvider
-
-    def get_version_status(self) -> VersionStatus:
-        aws_version = None
-        try:
-            response = subprocess.run("aws --version", capture_output=True, shell=True)
-            matched = re.match(r"aws-cli/([0-9.]+)", response.stdout.decode("utf8"))
-            aws_version = SemanticVersion.from_string(matched.group(1))
-        except ValueError:
-            pass
-
-        return VersionStatus(
-            aws_version, self.latest_version_provider.get_semantic_version("aws/aws-cli", True)
-        )
-
-
-class CopilotVersioning:
-    def __init__(self, latest_version_provider: VersionProvider = None):
-        self.latest_version_provider = latest_version_provider or GithubLatestVersionProvider
-
-    def get_version_status(self) -> VersionStatus:
-        copilot_version = None
-
-        try:
-            response = subprocess.run("copilot --version", capture_output=True, shell=True)
-            [copilot_version] = re.findall(r"[0-9.]+", response.stdout.decode("utf8"))
-        except ValueError:
-            pass
-
-        return VersionStatus(
-            SemanticVersion.from_string(copilot_version),
-            self.latest_version_provider.get_semantic_version("aws/copilot-cli"),
-        )
